// src/models/types/fridge.ts
export enum FridgeItemPriority {
  LOW = 'LOW',
  MEDIUM = 'MEDIUM',
  HIGH = 'HIGH',
  VERY_HIGH = 'VERY_HIGH',
}

export enum FridgeItemPriorityValue {
  LOW = 'Thấp',
  MEDIUM = 'Trung bình',
  HIGH = 'Cao',
  VERY_HIGH = 'Rất cao',
}

export enum FridgeItemSortBy {
  CREATED_AT = 'CREATED_AT',
  DUE_DATE = 'DUE_DATE',
  PRIORITY = 'PRIORITY',
  QUANTITY = 'QUANTITY',
}